requirements.txt
# Django & REST Framework
Django>=4.2.0
djangorestframework==3.14.0

# CORS Support
django-cors-headers==4.0.0

# Database
PyMySQL==1.0.2

# Utilities
python-dotenv==1.0.0

# Authentication
djoser==2.3.3
djangorestframework-simplejwt==5.5.1
Pillow==9.4.0
cloudinary==1.44.2
django-cloudinary-storage==0.3.0